+++ Data/DataAbstractAPI.cs
@@ -1,10 +1,17 @@
-﻿namespace Data
+﻿using System.ComponentModel;
+
+namespace Data
 {
     public abstract class DataAbstractAPI
     {
         public static DataAbstractAPI CreateApi()
         {
-            return new DataAPI();
+            return new PoolTable(500, 300);
         }
+        public abstract int Width { get; }
+        public abstract int Height { get; }
+        public abstract void CreateBalls(int ballsQuantity, int radius);
+        public abstract List<IBall> GetAllBalls();
+        public abstract void RemoveBalls();
     }
 }
